add Header component to all pages, containig header
