+++ project/src/components/cities-list/cities-list.jsx
@@ -44,5 +44,5 @@ const mapDispatchToProps = (dispatch) => ({
   },
 });
 
-export {CitiesList};
+export { CitiesList };
 export default connect(mapStateToProps, mapDispatchToProps)(CitiesList);
