+++ project/src/components/main-page/main-page.jsx
@@ -2,8 +2,6 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import OffersList from '../offers-list/offers-list';
 import { cardProp } from '../card/card.prop';
-import { Link } from 'react-router-dom';
-import { AppRoute } from '../../const';
 import Map from '../map/map';
 import { CardSetting } from '../../const';
 import { cities } from '../../const';
@@ -13,11 +11,11 @@ import OffersSorting from '../offers-sorting/offers-sorting';
 import { getFilteredOffers } from '../../utils/utils';
 import { sortOffers } from '../../utils/utils';
 import { SortingType } from '../../const';
-import { AuthorizationStatus } from '../../const';
-import { logout } from '../../store/api-actions';
+
+import Header from '../header/header';
 
 function MainPage(props) {
-  const {filteredOffers, activeCity, authorizationStatus, logoutApp} = props;
+  const {filteredOffers, activeCity} = props;
 
   const [activeCardId, setActiveCardId] = useState(null);
   const [activeSorting, setActiveSorting] = useState(SortingType.POPULAR);
@@ -68,58 +66,9 @@ function MainPage(props) {
     </div>
   );
 
-  const authorizedUserLink = (
-    <>
-      <li className="header__nav-item user">
-        <Link to={AppRoute.FAVORITES} className="header__nav-link header__nav-link--profile">
-          <div className="header__avatar-wrapper user__avatar-wrapper">
-          </div>
-          <span className="header__user-name user__name">{localStorage.getItem('login')}</span>
-        </Link>
-      </li>
-      <li className="header__nav-item">
-        <Link
-          onClick={(evt) => {
-            evt.preventDefault();
-            logoutApp();
-          }}
-          to='/'
-          className="header__nav-link"
-        >
-          <span className="header__signout">Sign out</span>
-        </Link>
-      </li>
-    </>
-  );
-
-  const unAthorizedUserLink = (
-    <li className="header__nav-item user">
-      <Link to={AppRoute.LOGIN} className="header__nav-link header__nav-link--profile">
-        <div className="header__avatar-wrapper user__avatar-wrapper">
-        </div>
-        <span className="header__login">Sign in</span>
-      </Link>
-    </li>
-  );
-
   return (
     <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link to={AppRoute.ROOT} className="header__logo-link header__logo-link--active">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                { authorizationStatus === AuthorizationStatus.AUTH ? authorizedUserLink : unAthorizedUserLink}
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <Header />
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
@@ -136,21 +85,12 @@ function MainPage(props) {
 MainPage.propTypes = {
   filteredOffers: PropTypes.arrayOf(cardProp.offer).isRequired,
   activeCity: PropTypes.string.isRequired,
-  authorizationStatus: PropTypes.string.isRequired,
-  logoutApp: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = (state) => ({
   filteredOffers: getFilteredOffers(state.offers, state.city),
   activeCity: state.city,
-  authorizationStatus: state.authorizationStatus,
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  logoutApp() {
-    dispatch(logout());
-  },
 });
 
 export { MainPage };
-export default connect(mapStateToProps, mapDispatchToProps)(MainPage);
+export default connect(mapStateToProps)(MainPage);
