+++ project/src/components/room/room.jsx
@@ -3,8 +3,6 @@ import CommentForm from '../comment-form/comment-form';
 import NotFound from '../not-found/not-found';
 import { defineRatingWidth } from '../../utils/utils';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
-import { AppRoute } from '../../const';
 import { cardProp } from '../card/card.prop';
 import ReviewsList from '../review-list/review-list';
 import Map from '../map/map';
@@ -12,6 +10,7 @@ import OffersList from '../offers-list/offers-list';
 import { roomTypeAlias } from '../../const';
 import { CardSetting } from '../../const';
 import { connect } from 'react-redux';
+import Header from '../header/header';
 
 function Room(props) {
   const {offers} = props;
@@ -31,34 +30,7 @@ function Room(props) {
 
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link to={AppRoute.ROOT} className="header__logo-link">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link to={AppRoute.FAVORITES} className="header__nav-link header__nav-link--profile">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="foo">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <Header />
       <main className="page__main page__main--property">
         <section className="property">
           <div className="property__gallery-container container">
