+++ project/src/components/sign-in/sign-in.jsx
@@ -110,13 +110,6 @@ function SignIn() {
               </button>
             </form>
           </section>
-          <section className="locations locations--login locations--current">
-            <div className="locations__item">
-              <a className="locations__item-link" href="foo">
-                <span>Amsterdam</span>
-              </a>
-            </div>
-          </section>
         </div>
       </main>
     </div>
